Added Authorization

+++ backend/routes/posts.js
@@ -32,8 +32,10 @@ router.post('', checkAuth,
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
+
   post.save().then((createdPost) => {
      res.status(201).json({
       message: 'Post added successfully.',
@@ -62,11 +64,16 @@ multer({storage: storage}).single("image"),
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then( result => {
-    res.status(200).json({ message: 'Update successful!'});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then( result => {
+    // if the post was modified
+    if (result.nModified > 0) { //how many entries were modified
+      res.status(200).json({ message: 'Update successful!'});
+    } else {
+      res.status(401).json({ message: 'Not Authorized!'});
+    }
   });
 });
 
@@ -107,9 +114,13 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth , (req, res, next) => {
-  Post.deleteOne({_id: req.params.id }).then( result => {
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId }).then( result => {
     console.log(result);
-    res.status(200).json({ message: 'Post deleted'});
+    if (result.deletedCount > 0) { // or result.n > 0 but I prefer to use deletedCount
+      res.status(200).json({ message: 'Deletion successful!'});
+    } else {
+      res.status(401).json({ message: 'Not Authorized!'});
+    }
   });
 });
